requirements.txt
Pillow
psycopg
psycopg2
psycopg-binary
bcrypt
reportlab
dlib
